pyproject.toml
[tool.poetry]
name = "drf_api"
version = "0.1.0"
description = ""
authors = ["monaSalih <monaalobisat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
flack = "^1.3.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
